requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
temporalio==1.4.0
redis==5.0.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
minio==7.2.4
python-multipart==0.0.9
httpx==0.26.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
open-telemetry-sdk==0.43b0
open-telemetry-exporter-otlp==0.43b0
prometheus-client==0.19.0
tritonclient[all]==2.41.1
numpy==1.26.4
pillow==10.2.0
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.21
unstructured==0.12.3
tika==2.6.0
opencv-python-headless==4.9.0.80
torch==2.2.0
transformers==4.37.2
triton-python-backend-utils==0.0.1
